requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==20.1.0
whitenoise==6.2.0
python-dotenv==1.0.1